Final Modifications

+++ GraduateProject/httpModels/api/response/TaskData.cs
@@ -7,5 +7,5 @@ public class TaskData
     public int TaskID { get; set; }
     public DateTime TaskDate { get; set; }
     public CurrentState CurrentState { get; set; }
-    public int result { get; set; }
+    public double result { get; set; }
 }
